+++ components/src/main/java/com/mercadolibre/android/andesui/datepicker/AndesDatePicker.kt
@@ -1,6 +1,7 @@
 package com.mercadolibre.android.andesui.datepicker
 
 import android.content.Context
+import android.os.Build
 import androidx.constraintlayout.widget.ConstraintLayout
 import android.util.AttributeSet
 import android.view.LayoutInflater
@@ -123,6 +124,11 @@ class AndesDatePicker : ConstraintLayout {
     private fun initComponents() {
         LayoutInflater.from(context).inflate(R.layout.andes_layout_datepicker, this)
         onCheckedChangeListener(andesBtnSelectDate)
+        val version: Int = Build.VERSION.SDK_INT
+        if (version <= API_LEVEL){
+            calendarView.layoutParams.width = WIDTH
+            calendarView.layoutParams.height = HEIGHT
+        }
     }
 
     /**
@@ -249,5 +255,8 @@ class AndesDatePicker : ConstraintLayout {
         private const val DEFAULT_MIN_DATE = -2208973392000
         private const val DEFAULT_MAX_DATE = 4133905200000
         private const val DATE_FORMAT = "dd/MM/yyyy"
+        private const val API_LEVEL = 21
+        private const val HEIGHT = 900
+        private const val WIDTH = 900
     }
 }